# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-SOC Cloud API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List all detection zones
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
    post:
      operationId: createZone
      tags: [zones]
      summary: Create a detection zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneInput"
      responses:
        "201":
          description: Zone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
  /zones/{id}:
    patch:
      operationId: updateZone
      tags: [zones]
      summary: Update detection zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneUpdate"
      responses:
        "200":
          description: Updated zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
    delete:
      operationId: deleteZone
      tags: [zones]
      summary: Delete zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Events ─────────────────────────────────────────────────────────────────
...
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incident reports
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create incident report from events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"

  # ── Dashboard / Analytics ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Zone:
      type: object
      required: [id, name, cameraId, type, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        cameraId: { type: integer }
        type: { type: string, enum: [restricted, monitored, safe] }
        coordinates: { type: ["string", "null"] }
        active: { type: boolean }
        createdAt: { type: string }
    ZoneInput:
      type: object
      required: [name, cameraId, type]
      properties:
        name: { type: string, minLength: 1 }
        cameraId: { type: integer }
        type: { type: string, enum: [restricted, monitored, safe] }
        coordinates: { type: string }
        active: { type: boolean }
    ZoneUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string, enum: [restricted, monitored, safe] }
        coordinates: { type: string }
        active: { type: boolean }

    # ── Security Event ────────────────────────────────────────────────────────
...
    Incident:
      type: object
      required: [id, title, severity, status, eventCount, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        severity: { type: string, enum: [low, medium, high, critical] }
        status: { type: string, enum: [open, investigating, resolved, closed] }
        eventCount: { type: integer }
        cameraIds: { type: ["string", "null"] }
        aiSummary: { type: ["string", "null"] }
        resolvedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    IncidentInput:
      type: object
      required: [title, severity]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
        eventIds: { type: string }
        cameraIds: { type: string }
    IncidentUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        severity: { type: string, enum: [low, medium, high, critical] }
        status: { type: string, enum: [open, investigating, resolved, closed] }
        aiSummary: { type: string }

    # ── Dashboard ─────────────────────────────────────────────────────────────
...