# lib/api-zod/src/generated/types/zoneInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI-SOC Cloud API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ZoneInputType } from './zoneInputType';

export interface ZoneInput {
  /** @minLength 1 */
  name: string;
  cameraId: number;
  type: ZoneInputType;
  coordinates?: string;
  active?: boolean;
}